requirements.txt
Flask == 2.2.2   
mysql.connector == 2.2.9